spike/pyproject.toml
[project]
name = "spike"
version = "0.0.0"
description = "1-day spike: validate ADK 2.0 Workflow + RequestInput + Memory Bank + Agent Runtime deploy"
requires-python = ">=3.12"
dependencies = [
    "google-adk==2.0.0b1",
    "google-cloud-aiplatform>=1.105",
    "pydantic>=2.0",
]
